refactor error parsing into its own class
